requirements.txt
stups-zign
requests
